+++ src/models/SettingsModel.php
@@ -499,9 +499,9 @@ class SettingsModel extends Model
     /**
      * Returns whether the cache should be cleared on refresh.
      *
-     * @since 4.0.0
+     * @since 4.14.0
      */
-    public function clearOnRefresh(bool $force = false): bool
+    public function shouldClearOnRefresh(bool $force = false): bool
     {
         if ($force) {
             return true;
@@ -512,11 +512,22 @@ class SettingsModel extends Model
     }
 
     /**
+     * Returns whether the cache should be cleared on refresh.
+     *
+     * @since 4.0.0
+     * @deprecated in 4.14.0. Use [[shouldClearOnRefresh()]] instead.
+     */
+    public function clearOnRefresh(bool $force = false): bool
+    {
+        return $this->shouldClearOnRefresh($force);
+    }
+
+    /**
      * Returns whether the cache should be expired on refresh.
      *
-     * @since 4.8.0
+     * @since 4.14.0
      */
-    public function expireOnRefresh(bool $forceClear = false, bool $forceGenerate = false): bool
+    public function shouldExpireOnRefresh(bool $forceClear = false, bool $forceGenerate = false): bool
     {
         if ($forceClear || $forceGenerate) {
             return false;
@@ -531,11 +542,22 @@ class SettingsModel extends Model
     }
 
     /**
+     * Returns whether the cache should be expired on refresh.
+     *
+     * @since 4.8.0
+     * @deprecated in 4.14.0. Use [[shouldExpireOnRefresh()]] instead.
+     */
+    public function expireOnRefresh(bool $forceClear = false, bool $forceGenerate = false): bool
+    {
+        return $this->shouldExpireOnRefresh($forceClear, $forceGenerate);
+    }
+
+    /**
      * Returns whether the cache should be generated on refresh.
      *
-     * @since 4.0.0
+     * @since 4.14.0
      */
-    public function generateOnRefresh(bool $force = false): bool
+    public function shouldGenerateOnRefresh(bool $force = false): bool
     {
         if ($force) {
             return true;
@@ -550,22 +572,44 @@ class SettingsModel extends Model
     }
 
     /**
+     * Returns whether the cache should be generated on refresh.
+     *
+     * @since 4.0.0
+     * @deprecated in 4.14.0. Use [[shouldGenerateOnRefresh()]] instead.
+     */
+    public function generateOnRefresh(bool $force = false): bool
+    {
+        return $this->shouldGenerateOnRefresh($force);
+    }
+
+    /**
      * Returns whether the cache should be purged after being refreshed.
      *
      * @since 4.8.0
      */
+    public function shouldPurgeAfterRefresh(bool $forceClear = false): bool
+    {
+        return $this->shouldExpireOnRefresh($forceClear);
+    }
+
+    /**
+     * Returns whether the cache should be purged after being refreshed.
+     *
+     * @since 4.8.0
+     * @deprecated in 4.14.0. Use [[shouldPurgeAfterRefresh()]] instead.
+     */
     public function purgeAfterRefresh(bool $forceClear = false): bool
     {
-        return $this->expireOnRefresh($forceClear);
+        return $this->shouldPurgeAfterRefresh($forceClear);
     }
 
     /**
      * Returns whether the page should be generated based on whether a query
      * string exists in the URI.
      *
-     * @since 4.4.0
+     * @since 4.14.0
      */
-    public function generatePageBasedOnQueryString(string $uri): bool
+    public function shouldGeneratePageBasedOnQueryString(string $uri): bool
     {
         if ($this->generatePagesWithQueryStringParams === true) {
             return true;
@@ -580,13 +624,36 @@ class SettingsModel extends Model
     }
 
     /**
+     * Returns whether the page should be generated based on whether a query
+     * string exists in the URI.
+     *
+     * @since 4.4.0
+     * @deprecated in 4.14.0. Use [[shouldGeneratePageBasedOnQueryString()]] instead.
+     */
+    public function generatePageBasedOnQueryString(string $uri): bool
+    {
+        return $this->shouldGeneratePageBasedOnQueryString($uri);
+    }
+
+    /**
+     * Returns whether the cache should be purged after being generated.
+     *
+     * @since 4.14.0
+     */
+    public function shouldPurgeAssetImages(): bool
+    {
+        return $this->purgeAssetImagesWhenChanged && $this->cachePurgerType !== DummyPurger::class;
+    }
+
+    /**
      * Returns whether the cache should be purged after being generated.
      *
      * @since 4.4.0
+     * @deprecated in 4.14.0. Use [[shouldPurgeAssetImages()]] instead.
      */
     public function purgeAssetImages(): bool
     {
-        return $this->purgeAssetImagesWhenChanged && $this->cachePurgerType !== DummyPurger::class;
+        return $this->shouldPurgeAssetImages();
     }
 
     /**